+++ src/components/atoms/InputLabel/InputLabel.tsx
@@ -1,21 +1,24 @@
-import React, { ReactNode, FC } from 'react';
+import React, { HTMLAttributes, FC } from 'react';
 import styled from 'styled-components';
-import Text from '../Text/Text';
+import { typography } from '../../../constants/typography';
 import { colors } from '../../../constants/colors';
 
-export interface IInputLabelProps {
+export interface IInputLabelProps extends HTMLAttributes<HTMLLabelElement> {
+  /**
+   * Define the associated input identifier
+   */
   htmlFor: string;
-  children: ReactNode;
 }
 
-const InputLabel = styled(Text).attrs({
-  forwardedAs: 'label',
-  weight: 'semibold',
-  color: colors.neutral.dark,
-  size: 'base',
-  mb: false,
-})<IInputLabelProps>`
+const InputLabel = styled.label<IInputLabelProps>`
   display: block;
+  margin: 0 0 10px;
+  letter-spacing: 0;
+  color: ${colors.neutral.dark};
+  line-height: ${typography.lineHeights.text};
+  font-family: ${typography.primary};
+  font-weight: ${typography.weights.semibold};
+  font-size: ${typography.sizes.text['base']};
 `;
 
 // TODO: Styleguidist to be able to locate styled components. See #147.