+++ src/components/styles/GlobalStyles.tsx
@@ -1,6 +1,7 @@
 import { createGlobalStyle, css } from 'styled-components';
 import { normalize } from 'styled-normalize';
 import { typography } from '../../constants/typography';
+import { version as tippyVersion } from 'tippy.js/package.json';
 
 // Universal box sizing with Inheritance. More info: https://css-tricks.com/box-sizing/#article-header-id-6
 const boxSizing = css`
@@ -28,11 +29,17 @@ const imagery = css`
   }
 `;
 
+const tooltipGlobalStyles = css`
+  @import url('https://unpkg.com/tippy.js@${tippyVersion}/dist/tippy.css');
+  @import url('https://unpkg.com/tippy.js@${tippyVersion}/animations/shift-toward.css');
+`;
+
 const GlobalStyles = createGlobalStyle`
   ${normalize}
   ${defaultFont}
   ${boxSizing}
   ${imagery}
+  ${tooltipGlobalStyles}
 `;
 
 export default GlobalStyles;