+++ src/index.ts
@@ -2,12 +2,14 @@
     Expose shared endpoints:
  */
 
-// Atoms
-// Constants (colors, fonts, ...)
-import { colors } from './constants/colors';
+// Global (colors, fonts, ...)
+export { typography } from './constants/typography';
+export { colors } from './constants/colors';
 export { default as grid } from './constants/grid';
-import { typography } from './constants/typography';
+export { default as Fonts } from './components/styles/Fonts';
+export { default as GlobalStyles } from './components/styles/GlobalStyles';
 
+// Atoms
 export { default as AlertBox } from './components/atoms/AlertBox/AlertBox';
 export { default as Badge } from './components/atoms/Badge/Badge';
 export { default as Button } from './components/atoms/Button/Button';
@@ -37,9 +39,9 @@ export { default as TextField } from './components/molecules/TextField/TextField
 export { default as DropdownField } from './components/molecules/DropdownField/DropdownField';
 
 // Organisms
-export { default as Form } from './components/organisms/Form';
+export * from './components/organisms/Form';
+export * from './components/organisms/Accordion';
 export { default as Navbar } from './components/organisms/Navbar';
-export { default as Accordion } from './components/organisms/Accordion';
 
 // Layout
 export { default as FlexContainer } from './components/layout/FlexContainer/FlexContainer';
@@ -55,14 +57,6 @@ export { default as ChevronIcon } from './components/icons/Chevron/Chevron';
 export { default as ZopaIcon } from './components/icons/ZopaLogo/ZopaLogo';
 export { default as HamburgerIcon } from './components/icons/Hamburger/Hamburger';
 export { default as ProfileIcon } from './components/icons/Profile/Profile';
-
-// Hooks
-export { default as useAccordion } from './components/hooks/useAccordion/useAccordion';
-export { default as useForm } from './components/hooks/useForm/useForm';
-
-export { colors, typography };
-export { default as Fonts } from './components/styles/Fonts';
-export { default as GlobalStyles } from './components/styles/GlobalStyles';
 
 // Helpers
 export * from './helpers/responsiveness';